.generator/requirements.txt
trame
pyyaml
black